+++ packages/hydrogen/src/utilities/log/log.ts
@@ -2,16 +2,22 @@ import {ServerComponentRequest} from '../../framework/Hydration/ServerComponentR
 import kleur from 'kleur';
 import {getTime} from '../timing';
 
-export interface RequestLogger extends Logger {
-  (request: ServerComponentRequest): Logger;
-}
-
 export interface Logger {
   trace: (...args: Array<any>) => void;
   debug: (...args: Array<any>) => void;
   warn: (...args: Array<any>) => void;
   error: (...args: Array<any>) => void;
   fatal: (...args: Array<any>) => void;
+}
+
+export function getLoggerFromContext(context: any): Logger {
+  return {
+    trace: (...args) => (globalThis as any).__hlogger.trace(context, ...args),
+    debug: (...args) => (globalThis as any).__hlogger.debug(context, ...args),
+    warn: (...args) => (globalThis as any).__hlogger.warn(context, ...args),
+    error: (...args) => (globalThis as any).__hlogger.error(context, ...args),
+    fatal: (...args) => (globalThis as any).__hlogger.fatal(context, ...args),
+  };
 }
 
 // @todo - multiple instances of log.ts are loaded, we utilitze the
@@ -38,22 +44,28 @@ export function setLogger(_logger: Logger) {
   (globalThis as any).__hlogger = _logger;
 }
 
-export const log: RequestLogger = function (
-  request: ServerComponentRequest
-): Logger {
-  return {
-    trace: (...args) => (globalThis as any).__hlogger.trace({request}, ...args),
-    debug: (...args) => (globalThis as any).__hlogger.debug({request}, ...args),
-    warn: (...args) => (globalThis as any).__hlogger.warn({request}, ...args),
-    error: (...args) => (globalThis as any).__hlogger.error({request}, ...args),
-    fatal: (...args) => (globalThis as any).__hlogger.fatal({request}, ...args),
-  };
+/**
+ * A utility for logging debugging, warning, and error information about the application.
+ * Use by importing `log` `@shopify/hydrogen` or by using a `log` prop passed to each page
+ * component. Using the latter is ideal, because it will ty your log to the current request in progress.
+ */
+export const log: Logger = {
+  trace(...args) {
+    return (globalThis as any).__hlogger.trace({}, ...args);
+  },
+  debug(...args) {
+    return (globalThis as any).__hlogger.debug({}, ...args);
+  },
+  warn(...args) {
+    return (globalThis as any).__hlogger.warn({}, ...args);
+  },
+  error(...args) {
+    return (globalThis as any).__hlogger.error({}, ...args);
+  },
+  fatal(...args) {
+    return (globalThis as any).__hlogger.trace({}, ...args);
+  },
 };
-log.trace = (...args) => (globalThis as any).__hlogger.trace({}, ...args);
-log.debug = (...args) => (globalThis as any).__hlogger.debug({}, ...args);
-log.warn = (...args) => (globalThis as any).__hlogger.warn({}, ...args);
-log.error = (...args) => (globalThis as any).__hlogger.error({}, ...args);
-log.fatal = (...args) => (globalThis as any).__hlogger.trace({}, ...args);
 
 export function logServerResponse(
   type: 'str' | 'rsc' | 'ssr',