+++ packages/hydrogen/CHANGELOG.md
@@ -13,6 +13,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - feat: upgrade to latest React 18 experimental version
 - docs: product provider tweaks
 - docs: Document naming conventions
+- feat: logging abstraction with default timing information
 
 ## 0.8.0 - 2021-12-07